Move Credentials to sessions
